pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "topstep-quant"
version = "0.1.0"
description = "Autonomous trading bot for CME micro-futures in a Topstep 50K account, with strict risk controls."
authors = ["Your Name <[email protected]>"]
readme = "README.md"
license = "MIT"
keywords = ["trading", "algorithmic-trading", "futures", "topstep", "quant"]
homepage = "https://github.com/yourusername/topstep_quant"
packages = [{include = "topstep_quant"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.3.0"
pandas = "^2.3.0"
requests = "^2.32.4"
prometheus-client = "^0.22.1"
pyyaml = "^6.0"
jsonschema = "^4.17.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.10.0"
black = "^25.1.0"
mypy = "^1.15.0"
bandit = "^1.8.3"
pytest = "^8.4.0"

[tool.poetry.scripts]
topstep-quant = "topstep_quant.bot:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "N", "UP", "ANN", "S", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["ANN101", "ANN102", "S101", "PLR0913", "PLR0912", "PLR0915"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]